feat(ui): unify clipboard copy with shared hook and success feedback
