+++ frontend/public/assets/i18n/ru.json
@@ -1551,6 +1551,7 @@
   },
   "MESSAGES": {
     "CLIPBOARD_NOT_SUPPORTED": "Буфер обмена не поддерживается.\nВаш браузер или текущий контекст не позволяют доступ к буферу обмена.\nПожалуйста, используйте HTTPS или localhost.",
+    "COPIED_TO_CLIPBOARD": "Скопировано в буфер обмена.",
     "CONFIG_CHANGED": "Конфигурация на сервере изменилась.",
     "CONFIRM_DELETE": "Вы действительно хотите удалить это?",
     "CONFIRM_SOURCES_SAVE": "Вы действительно хотите сохранить?",
