+++ frontend/src/hooks/mod.rs
@@ -1,9 +1,11 @@
+mod use_clipboard_copy;
 mod use_icon_context;
 mod use_key_down;
 mod use_server_status;
 mod use_service_context;
 mod use_websocket_status;
 
+pub use use_clipboard_copy::*;
 pub use use_icon_context::*;
 pub use use_key_down::*;
 pub use use_server_status::*;
